Represent byte, char and tree counts as 32 bit numbers
The parser spends the majority of its time allocating and freeing trees and stack nodes.
Also, the memory footprint of the AST is a significant concern when using tree-sitter
with large files. This library is already unlikely to work very well with source files
larger than 4GB, so representing rows, columns, byte lengths and child indices as
unsigned 32 bit integers seems like the right choice.

+++ include/tree_sitter/parser.h
@@ -58,8 +58,8 @@ typedef union {
 } TSParseActionEntry;
 
 typedef struct TSLanguage {
-  size_t symbol_count;
-  size_t token_count;
+  uint32_t symbol_count;
+  uint32_t token_count;
   const char **symbol_names;
   const TSSymbolMetadata *symbol_metadata;
   const unsigned short *parse_table;
