+++ spec/helpers/spy_input.cc
@@ -22,7 +22,7 @@ SpyInput::~SpyInput() {
   delete[] buffer;
 }
 
-const char * SpyInput::read(void *payload, size_t *bytes_read) {
+const char * SpyInput::read(void *payload, uint32_t *bytes_read) {
   auto spy = static_cast<SpyInput *>(payload);
 
   if (spy->byte_offset > spy->content.size()) {
@@ -52,7 +52,7 @@ const char * SpyInput::read(void *payload, size_t *bytes_read) {
   return spy->buffer;
 }
 
-int SpyInput::seek(void *payload, size_t character, size_t byte) {
+int SpyInput::seek(void *payload, uint32_t character, uint32_t byte) {
   auto spy = static_cast<SpyInput *>(payload);
   if (spy->strings_read.size() == 0 || spy->strings_read.back().size() > 0)
     spy->strings_read.push_back("");
